refactor: move KMSDecrypter interface to pkg/crypto
Move KMSDecrypter interface and AWSKMSDecrypter from pkg/store to
pkg/crypto where other crypto primitives live. Keep store-level
functions (OpenWithKMS, ExtractMasterKeyWithKMS) in pkg/store since
they depend on store.KeySlot.
